fix: rename files with typos
- SyntaxHightlightedTextarea.svelte → SyntaxHighlightedTextarea.svelte
- ToolConvertion.svelte → ToolConversion.svelte
- getRecomended.ts → getRecommended.ts
